pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "google-genai>=1.32.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-pptx>=0.6.23",
    "sift-stack-py>=0.8.4",
    "streamlit>=1.49.0",
    "textract>=1.6.5",
]
